[DF-9566] x64 ffmpeg libs. Compilation fixes.
Former-commit-id: ca734ceeeb0213bcb318dae98e7f766c4ce48e9b

+++ Sources/Internal/Platform/TemplateWin32/MovieViewControlFFMPEG.cpp
@@ -551,7 +551,7 @@ PixelFormat MovieViewControl::GetPixelFormat() const
 
 Vector2 MovieViewControl::GetResolution() const
 {
-    return Vector2(frameWidth, frameHeight);
+    return Vector2(static_cast<float32>(frameWidth), static_cast<float32>(frameHeight));
 }
 
 void MovieViewControl::UpdateDrawData(DecodedFrameBuffer* buffer)
